+++ custom_components/small_grow_tent_controller/const.py
@@ -1,5 +1,5 @@
 DOMAIN = "small_grow_tent_controller"
-VERSION = "0.1.24"
+VERSION = "0.1.25"
 
 PLATFORMS = ["sensor", "switch", "select", "number", "time", "binary_sensor", "button"]